Clear out the rest of the cpp files
The test suite definitely doesn't compile right now but it's all just
headers at this time.
