+++ rice/Exception.hpp
@@ -1,10 +1,6 @@
 #ifndef Rice__Exception__hpp_
 #define Rice__Exception__hpp_
 
-namespace Rice {
-  class Exception;
-}
-
 #include "Exception_defn.hpp"
 #include "Exception_Base.hpp"
 #include "Address_Registration_Guard.hpp"
